refactor(srf_parser): use 32-bit indexing instead of 64-bit indexing

+++ source_modelling/srf_parser/src/lib.rs
@@ -11,8 +11,8 @@ use std::io::Error;
 
 #[derive(Default)]
 struct SparseMatrix {
-    row_ptr: Vec<usize>,
-    col_ptr: Vec<usize>,
+    row_ptr: Vec<u32>,
+    col_ptr: Vec<u32>,
     data: Vec<f32>,
 }
 
@@ -102,7 +102,7 @@ fn read_srf_points(
         let slip1 = parse_value::<f32>(data, &mut index)?;
         metadata.push(slip1);
 
-        let nt = parse_value::<usize>(data, &mut index)?;
+        let nt = parse_value::<u32>(data, &mut index)?;
 
         let _nt2 = parse_value::<f32>(data, &mut index)?;
         let _slip2 = parse_value::<i64>(data, &mut index)?;
@@ -111,8 +111,8 @@ fn read_srf_points(
 
         metadata.push((nt as f32) * dt);
 
-        let start_column_index: usize = (tinit / dt).floor() as usize;
-        slipt1.row_ptr.push(slipt1.data.len());
+        let start_column_index: u32 = (tinit / dt).floor() as u32;
+        slipt1.row_ptr.push(slipt1.data.len() as u32);
         for i in start_column_index..start_column_index + nt {
             let slip = parse_value::<f32>(data, &mut index)?;
             slipt1.col_ptr.push(i);
